requirements.txt
django
pillow
wand